requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
cryptography
openai
google-generativeai
langchain
langchain-google-genai
langchain-openai
langchain-core
langchain-community
sentence-transformers
qdrant-client
pypdf
arize-phoenix
requests
python-dotenv
streamlit-aggrid
